Add generate of world.

+++ src/grid/segment/cell.rs
@@ -8,13 +8,13 @@ use super::Genome;
 /// `Cell` is the main, executive essence of the simulation.
 #[derive(Debug, Clone, Default)]
 pub struct Cell {
-    pub energy: u16,
+    pub energy: f32,
     pub genome: Genome,
 }
 
 impl Cell {
     /// Creates a new cell with a given amount of energy.
-    pub fn new(energy: u16) -> Self {
+    pub fn new(energy: f32) -> Self {
         Self {
             energy,
             genome: Default::default(),
